Upgrade to "split" Unity 5 packages
Unity 5 has an undeclared dependency on a CommonServiceLocator binary which is shipped with the package, and additionaly has now split Unity into multiple packages.

In order to avoid the possible issues caused by an undeclared dependency, this commit updates to the latest versions of the previous Unity dependencies, and updates the major version to reflect the breaking change.